dom/base/test/test_bug894874.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=894874
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 894874</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <!-- IMPORTANT: Keep the sheets in this order! -->
  <link rel="stylesheet" type="text/css" href="data:text/css,">
  <link rel="stylesheet" type="text/css" title="one" href="data:text/css,">
  <link rel="stylesheet" type="text/css" title="two" href="data:text/css,">
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug 894874 **/
  SimpleTest.waitForExplicitFinish();
  addLoadEvent(function() {
    is(document.styleSheets[0].disabled, false,
       "Sheet with no title should be enabled");
    is(document.styleSheets[1].disabled, false,
       "First preferred sheet should be enabled");
    is(document.styleSheets[2].disabled, true,
       "Second preferred sheet should be disabled");
    is(document.selectedStyleSheetSet, "one", "Sheet one is enabled");
    document.styleSheets[0].disabled = true;
    document.styleSheets[2].disabled = false;
    ok(document.selectedStyleSheetSet === null,
       "Sheet one and sheet two are both enabled");
    SimpleTest.finish();
  });

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=894874">Mozilla Bug 894874</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
